% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ChangePointsPlot}
\alias{ChangePointsPlot}
\title{Plot Peak Ranges of Change Points}
\usage{
ChangePointsPlot(y, result, ...)
}
\arguments{
\item{y}{The original data to find change points. Must be one dimensional
data.}

\item{result}{The result of function \emph{MultiWindow}.}

\item{...}{Arguments to be passed to plot, such as \emph{main, xlab, ylab}.}
}
\value{
A plot of original data and peak ranges of change points.
}
\description{
Plot the peak ranges of change points produced by \emph{MultiWindow}. 
The blue solid line is the start of a peak range and the red dashed line
is the end of that peak range.
}
\examples{
N <- 1000
N1 <- floor(0.1*N)
N2 <- floor(0.3*N)
a1 <- c(0.8, -0.3); c1 <- 0
a2 <- c(-0.5, 0.1); c2 <- 0
a3 <- c(0.5, -0.5); c3 <- 0
y <- rep(0,N)
L<-2
y[1:L] <- rnorm(L)
for (n in (L+1):N){
  if (n <= N1) {
    y[n] <- y[(n-1):(n-L)] \%*\% a1 + c1 + rnorm(1)
  } else if (n <= (N1+N2)) {
    y[n] <- y[(n-1):(n-L)] \%*\% a2 + c2 + rnorm(1)
  }
  else {
    y[n] <- y[(n-1):(n-L)] \%*\% a3 + c3 + rnorm(1)
  }
}
result <- MultiWindow(y,window_list=c(100,50,20,10,5),point_max=5)
ChangePointsPlot(y,result)
}
\references{
J. Ding, Y. Xiang, L. Shen, and V. Tarokh, \emph{Multiple Change
  Point Analysis: Fast Implementation and Strong Consistency}. IEEE
  Transactions on Signal Processing, vol. 65, no. 17, pp. 4495-4510, 2017.
}
