% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change_point.R
\name{GetLogLik}
\alias{GetLogLik}
\title{Get Log Likelihood}
\usage{
GetLogLik(y, left, right)
}
\arguments{
\item{y}{The data to calculate log likelihood. The data must be one dimesional.}

\item{left}{The left end of the data that users want to use.}

\item{right}{The right end of the data that users want to use.}
}
\value{
log_lik
}
\description{
For a series of one dimensional data, get the log likelihood.
}
